AI Movement and Base Needs Script
Added the basis of the movement for the NPC, as well as a basic Need class, and an NPC Behaviour script which makes the NPC behave like it should. I used a weight and priority system to the NPCs needs, making it prioritise tasks such as food and sleep more than job and socialise.